refactor: replace default password
pwgen is not included by default and must be installed separate. If the user wished to use a random password every time he should install and configure the package by himself. Also the user wouldn't be able to know the default password, since the password isn't logged anywhere for security reasons.